move Source definitions to a single folder without breaking the API
