+++ pkg/source/source.go
@@ -18,22 +18,14 @@ package source
 
 import (
 	"context"
-	"fmt"
-	"sync"
 
 	"k8s.io/client-go/util/workqueue"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
-	internal "sigs.k8s.io/controller-runtime/pkg/internal/source"
+	internal "sigs.k8s.io/controller-runtime/pkg/source/internal"
 
 	"sigs.k8s.io/controller-runtime/pkg/cache"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
-)
-
-const (
-	// defaultBufferSize is the default number of event notifications that can be buffered.
-	defaultBufferSize = 1024
 )
 
 // Source is a source of events (eh.g. Create, Update, Delete operations on Kubernetes Objects, Webhook callbacks, etc)
@@ -62,164 +54,19 @@ func Kind(cache cache.Cache, object client.Object) SyncingSource {
 	return &internal.Kind{Type: object, Cache: cache}
 }
 
-var _ Source = &Channel{}
+// Informer is used to provide a source of events originating inside the cluster from Watches (e.g. Pod Create).
+type Informer = internal.Informer
 
 // Channel is used to provide a source of events originating outside the cluster
 // (e.g. GitHub Webhook callback).  Channel requires the user to wire the external
 // source (eh.g. http handler) to write GenericEvents to the underlying channel.
-type Channel struct {
-	// once ensures the event distribution goroutine will be performed only once
-	once sync.Once
-
-	// Source is the source channel to fetch GenericEvents
-	Source <-chan event.GenericEvent
-
-	// dest is the destination channels of the added event handlers
-	dest []chan event.GenericEvent
-
-	// DestBufferSize is the specified buffer size of dest channels.
-	// Default to 1024 if not specified.
-	DestBufferSize int
-
-	// destLock is to ensure the destination channels are safely added/removed
-	destLock sync.Mutex
-}
-
-func (cs *Channel) String() string {
-	return fmt.Sprintf("channel source: %p", cs)
-}
-
-// Start implements Source and should only be called by the Controller.
-func (cs *Channel) Start(
-	ctx context.Context,
-	handler handler.EventHandler,
-	queue workqueue.RateLimitingInterface,
-	prct ...predicate.Predicate) error {
-	// Source should have been specified by the user.
-	if cs.Source == nil {
-		return fmt.Errorf("must specify Channel.Source")
-	}
-
-	// use default value if DestBufferSize not specified
-	if cs.DestBufferSize == 0 {
-		cs.DestBufferSize = defaultBufferSize
-	}
-
-	dst := make(chan event.GenericEvent, cs.DestBufferSize)
-
-	cs.destLock.Lock()
-	cs.dest = append(cs.dest, dst)
-	cs.destLock.Unlock()
-
-	cs.once.Do(func() {
-		// Distribute GenericEvents to all EventHandler / Queue pairs Watching this source
-		go cs.syncLoop(ctx)
-	})
-
-	go func() {
-		for evt := range dst {
-			shouldHandle := true
-			for _, p := range prct {
-				if !p.Generic(evt) {
-					shouldHandle = false
-					break
-				}
-			}
-
-			if shouldHandle {
-				func() {
-					ctx, cancel := context.WithCancel(ctx)
-					defer cancel()
-					handler.Generic(ctx, evt, queue)
-				}()
-			}
-		}
-	}()
-
-	return nil
-}
-
-func (cs *Channel) doStop() {
-	cs.destLock.Lock()
-	defer cs.destLock.Unlock()
-
-	for _, dst := range cs.dest {
-		close(dst)
-	}
-}
-
-func (cs *Channel) distribute(evt event.GenericEvent) {
-	cs.destLock.Lock()
-	defer cs.destLock.Unlock()
-
-	for _, dst := range cs.dest {
-		// We cannot make it under goroutine here, or we'll meet the
-		// race condition of writing message to closed channels.
-		// To avoid blocking, the dest channels are expected to be of
-		// proper buffer size. If we still see it blocked, then
-		// the controller is thought to be in an abnormal state.
-		dst <- evt
-	}
-}
-
-func (cs *Channel) syncLoop(ctx context.Context) {
-	for {
-		select {
-		case <-ctx.Done():
-			// Close destination channels
-			cs.doStop()
-			return
-		case evt, stillOpen := <-cs.Source:
-			if !stillOpen {
-				// if the source channel is closed, we're never gonna get
-				// anything more on it, so stop & bail
-				cs.doStop()
-				return
-			}
-			cs.distribute(evt)
-		}
-	}
-}
-
-// Informer is used to provide a source of events originating inside the cluster from Watches (e.g. Pod Create).
-type Informer struct {
-	// Informer is the controller-runtime Informer
-	Informer cache.Informer
-}
-
-var _ Source = &Informer{}
-
-// Start is internal and should be called only by the Controller to register an EventHandler with the Informer
-// to enqueue reconcile.Requests.
-func (is *Informer) Start(ctx context.Context, handler handler.EventHandler, queue workqueue.RateLimitingInterface,
-	prct ...predicate.Predicate) error {
-	// Informer should have been specified by the user.
-	if is.Informer == nil {
-		return fmt.Errorf("must specify Informer.Informer")
-	}
-
-	_, err := is.Informer.AddEventHandler(internal.NewEventHandler(ctx, queue, handler, prct).HandlerFuncs())
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (is *Informer) String() string {
-	return fmt.Sprintf("informer source: %p", is.Informer)
-}
-
-var _ Source = Func(nil)
+type Channel = internal.Channel
 
 // Func is a function that implements Source.
-type Func func(context.Context, handler.EventHandler, workqueue.RateLimitingInterface, ...predicate.Predicate) error
+type Func = internal.Func
 
-// Start implements Source.
-func (f Func) Start(ctx context.Context, evt handler.EventHandler, queue workqueue.RateLimitingInterface,
-	pr ...predicate.Predicate) error {
-	return f(ctx, evt, queue, pr...)
-}
-
-func (f Func) String() string {
-	return fmt.Sprintf("func source: %p", f)
-}
+var _ Source = &internal.Kind{}
+var _ SyncingSource = &internal.Kind{}
+var _ Source = &internal.Informer{}
+var _ Source = &internal.Channel{}
+var _ Source = internal.Func(nil)
